breadcrumbs-refactor: created base skeleton for breadcrumbs pages and refactored breadcrumbs, implemented breadcrumbs provider
