+++ apps/console/src/app/(protected)/risks/create/page.tsx
@@ -1,5 +1,5 @@
 'use client'
-import React from 'react'
+import React, { useEffect } from 'react'
 import { PageHeading } from '@repo/ui/page-heading'
 import { useSession } from 'next-auth/react'
 import ProtectedArea from '@/components/shared/protected-area/protected-area.tsx'
@@ -7,10 +7,19 @@ import { useOrganizationRole } from '@/lib/authz/access-api.ts'
 import { canCreate } from '@/lib/authz/utils.ts'
 import { AccessEnum } from '@/lib/authz/enums/access-enum.ts'
 import CreateRiskForm from '@/components/pages/protected/risks/create/form/create-risk-form.tsx'
+import { BreadcrumbContext } from '@/providers/BreadcrumbContext.tsx'
 
 const Page: React.FC = () => {
   const { data: session } = useSession()
   const { data: permission, isLoading } = useOrganizationRole(session)
+  const { setCrumbs } = React.useContext(BreadcrumbContext)
+
+  useEffect(() => {
+    setCrumbs([
+      { label: 'Home', href: '/dashboard' },
+      { label: 'Risks', href: '/risks' },
+    ])
+  }, [setCrumbs])
 
   return (
     <>
